requirements.txt
mouse
keyboard
pyautogui
opencv-python
python-dotenv